Add loginUser function and update return types
The commit introduces a new loginUser function in UserService for user authentication, accepting UserLoginDTO as an argument. It also updates the return type of createUser and update functions from UserModel to UserResponseDTO for better data encapsulation. A user's details are now returned in a UserResponseDTO object instead of a UserModel object.