Add NewsItem model and refactor NewsService and NewsController classes
Introduced the NewsItem model and refactored the NewsController and NewsService classes to effectively retrieve the user's latest news. The previous implementation used Mono to manage the response from the news service which has been replaced with Flux to better handle multiple NewsItems. Also, a clearer separation of responsibilities in the code has been achieved with these modifications.