Relocate model classes to global.model package
Moved several exception and service classes from the "model" package to a new "global.model" package. This relocation clarifies the global scope of these classes and improves the project structure for better understanding and readability. The changes have also been reflected in all the classes that use these moved classes.