Introduce UserNotFoundException to InstagramTokenService
This update refactors the logic of the getAccessToken function in InstagramTokenService. Instead of directly trying to find the access token, it now first checks if the user exists. If the user does exist, only then will it attempt to find the access token. This change will help track issues related to non-existing users.