Update UnsplashAuthDetails with recentPhotos attribute
The UnsplashAuthDetails class in the user model is updated with a new attribute 'recentPhotos'. This attribute is a mutable list that will hold Photo objects, providing a way to keep a record of the recent photos associated with each Unsplash user. This enriches our user data model for improved functionality and user experience.