Add logging to various service methods and controllers
This commit introduces logging to various methods in the service and controller classes. The logs aid in debugging by providing clarity on the execution flow of the application. LoggerFactory from the SLF4J framework was utilized to create the logger instances, which have been added to classes such as SpotifyApiService, AESEncryption, TodoistTokenService and many more.