Refactor directory hierarchy for calendar and news services
The directories for the news and calendar services have been restructured to better reflect the different technologies utilized (Nextcloud and CalDav). Files have been moved to subfolders named after the technically corresponding service. The update covers multiple controller and service files in the news and calendar modules, enhancing the project structure for improved readability and manageability.