Refactor Todoist-related services and their corresponding classes
This commit introduces modifications in the directory structure and class names related to Todoist service. The refactoring primarily focuses on moving the authentication-related classes to a directory named 'auth', under 'widgets'. Moreover, the controllers and services specifically dealing with the tasks are now situated in 'tasks' directory, a subdirectory of 'widgets'. These changes aim to better organize the project structure, hence easing navigation and maintenance.