Add findById and save methods to UserService
Added a 'findById' and 'save' method to UserService.kt to increase its functionality. The 'findById' method is used to find a user by their ID whilst the 'save' method is used to store a user's information. In case of any error in saving the data, the 'save' method throws a 'CannotSaveUserException'.