Remove AESEncryption from EventService
The AESEncryption dependency was removed from EventService class and the decryptions in the CalendarService were rearranged. Logging messages were updated for better clarity on the actions being performed, especially in the getEvents and getCalendarEvents methods. The way WebClient was built has also been changed to enhance readility of the code.