Add debug statements to CalDav services
Three different debug log statements have been added in this commit. The debug statements are inserted specifically in the NextcloudCalDavService and CalDavEntityService methods. They provide logging information when getting remote calendars' raw data, filtering active and future entities, and mapping entities to CalDavEntity. This will be useful for monitoring and troubleshooting the application.