Relocate StateValidation and StateParameter to user module
Moved StateValidation class and StateParameter model from the global scope to the user service. This change reflects the more user-related functionality of these objects and hence improves the project structure. Modified all imports of these objects in various authentication services to reflect the changes. The annotation for StateValidation was switched from `@Component` to `@Service` to suit its new location.