Refactor CalDavService to handle duplicate resources
Changes have been introduced to the `CalDavService` which handle the scenario when duplicate resources are encountered. The `checkForDuplicates` method has been modified to return a Mono<Boolean> and accordingly, the resource addition logic is modified. A clear error message will now be thrown if duplicates are found. Moreover, decrypted resources are now returned in the `getUserResources` method.